+++ src/test/java/net/mureng/mureng/question/web/QuestionControllerTest.java
@@ -33,9 +33,6 @@ public class QuestionControllerTest extends AbstractControllerTest {
     @MockBean
     private QuestionService questionService;
 
-    @MockBean
-    private ReplyService replyService;
-
     private static final Long QUESTION_ID = 1L;
 
     @Nested
@@ -138,32 +135,6 @@ public class QuestionControllerTest extends AbstractControllerTest {
                 .andExpect(MockMvcResultMatchers.jsonPath("$.data[0].repliesCount").value(2))
                 .andExpect(MockMvcResultMatchers.jsonPath("$.data[1].content").value("This is english content."))
                 .andExpect(MockMvcResultMatchers.jsonPath("$.data[1].repliesCount").value(0))
-                .andDo(print());
-    }
-
-    @Test
-    @WithMockMurengUser
-    public void 질문_관련_답변_목록_가져오기_테스트() throws Exception {
-        Reply reply1 = EntityCreator.createReplyEntity();
-        reply1.setContent("content1");
-        Reply reply2 = EntityCreator.createReplyEntity();
-        reply2.setContent("content2");
-        reply2.setReplyLikes(new HashSet<>());
-        List<Reply> replies = Arrays.asList(reply1, reply2);
-        int page = 0;
-        int size = 2;
-
-        given(replyService.findRepliesByQuestionId(eq(1L), eq(PageRequest.of(page, size)), any()))
-                .willReturn(new PageImpl<>(replies, PageRequest.of(page, size), 2));
-
-        mockMvc.perform(
-                get("/api/questions/1/replies?page=0&size=2")
-        ).andExpect(status().isOk())
-                .andExpect(MockMvcResultMatchers.jsonPath("$.message").value("ok"))
-                .andExpect(MockMvcResultMatchers.jsonPath("$.data[0].content").value("content1"))
-                .andExpect(MockMvcResultMatchers.jsonPath("$.data[0].replyLikeCount").value(2))
-                .andExpect(MockMvcResultMatchers.jsonPath("$.data[1].content").value("content2"))
-                .andExpect(MockMvcResultMatchers.jsonPath("$.data[1].replyLikeCount").value(0))
                 .andDo(print());
     }
 }
